Delete original graphcast_simple script
