+++ examples/graphcast_simple.py
@@ -13,12 +13,37 @@ TODO
   the graphcast wrapper.
 """
 # %%
-from earth2mip.networks.graphcast import channels, inference
-from earth2mip.model_registry import Package
-import xarray
+import sys
+
+sys.path.insert(0, "..")
 import datetime
+
 import matplotlib.pyplot as plt
-from cartopy import crs
+import numpy as np
+import torch
+import xarray
+
+from earth2mip.model_registry import Package
+from earth2mip.networks.graphcast import channels, inference
+
+
+def get_input_from_xarray(task_config, example_batch):
+    arrays = []
+    levels = list(task_config.pressure_levels)
+    for v in task_config.target_variables:
+        if channels.is_3d(v):
+            # b, h, p, y, x
+            arr = example_batch[v].sel(level=levels).isel(time=slice(0, 2)).values
+            assert arr.ndim == 5
+            arrays.append(arr)
+        else:
+            arr = example_batch[v].isel(time=slice(0, 2)).values
+            assert arr.ndim == 4
+            arr = np.expand_dims(arr, 2)
+            arrays.append(arr)
+
+    return np.concatenate(arrays, axis=2)
+
 
 # %%
 
@@ -29,35 +54,30 @@ root = "/lustre/fsw/sw_earth2_ml/graphcast/"
 # https://console.cloud.google.com/storage/browser/dm_graphcast/dataset?pageState=(%22StorageObjectListTable%22:(%22f%22:%22%255B%255D%22))&prefix=&forceOnObjectsSortingFiltering=false
 root = "gs://dm_graphcast"
 package = Package(root, seperator="/")
-model = inference.Graphcast(package, device="cuda:0")
+time_loop = inference.load_time_loop(package)
 
-# %%
-# about 2 GB of data, so expect this to take some time
-# it will be locally cached, so you can evaluate it
 dataset_filename = package.get(
     "dataset/source-era5_date-2022-01-01_res-0.25_levels-37_steps-01.nc"
 )
 with open(dataset_filename, "rb") as f:
     example_batch = xarray.load_dataset(f).compute()
 
-packed = channels.pack(example_batch, model.codes)
-# only use first two input levels
-packed = packed[:, : model.n_history_levels]
-# graphcast uses -90 to 90 for lat, need to rev to put on our grid
-packed = packed[:, :, :, ::-1]
-time = datetime.datetime(2022, 1, 1)
-
 # %%
-i = model.out_channel_names.index("t2m")
-fig = plt.figure()
-for k, (t, y) in enumerate(model(time, packed)):
-    print(t)
-    plt.clf()
-    ax = fig.add_subplot(projection=crs.PlateCarree())
-    ax.pcolormesh(model.grid.lon, model.grid.lat, y[0, i], transform=crs.PlateCarree())
-    ax.coastlines(color="w")
-    plt.savefig(f"{k:04d}.png")
-    if k == 1:
+
+
+task_config = time_loop.task_config
+target_codes = channels.get_codes(
+    task_config.target_variables, task_config.pressure_levels, [0]
+)
+array = get_input_from_xarray(task_config, example_batch)
+pt = torch.from_numpy(array).cuda()
+
+time = datetime.datetime(2018, 1, 1)
+i = time_loop.out_channel_names.index("q925")
+
+for k, (time, x, _) in enumerate(time_loop(time, pt)):
+    print(k)
+    plt.pcolormesh(x[0, i].cpu().numpy())
+    plt.savefig(f"{k:03d}.png")
+    if k == 10:
         break
-
-# %%
